+++ internal/api/web/_app/immutable/entry/start.BhAMkR0X.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/DZKGm6l3.js";export{e as load_css,t as start};